backend/requirements.txt
asgiref==3.8.1
Django==4.2.21
django-cors-headers==4.7.0
django-filter==23.1
djangorestframework==3.15.1
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
packaging==25.0
pycodestyle==2.13.0
pyflakes==3.3.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
